// src/client/valibot.gen.ts
export const vDetailedErrorItem = v.object({
    field: v.optional(v.string()),
    message: v.string()
});
// ...
export const vProblemDetailsResponse = v.object({
    status: v.pipe(v.number(), v.integer(), v.minValue(-2147483648, 'Invalid value: Expected int32 to be >= -2147483648'), v.maxValue(2147483647, 'Invalid value: Expected int32 to be <= 2147483647')),
    title: v.string(),
    errors: v.optional(v.array(vDetailedErrorItem))
});